Original merge commit:
    commit 2e4a18e3e6f73c41d4935aed22330adf83f99f28 (HEAD)
    Merge: b91b6cab3a1 29f973a6162
    Author: Guillaume Damiand <guillaume.damiand@liris.cnrs.fr>
    Date:   Thu Mar 12 18:05:22 2020 +0100

        Merge master; NYF

+++ CGAL_ipelets/include/CGAL/CGAL_Ipelet_base_v7.h
@@ -1,20 +1,11 @@
 // Copyright (c) 2005-2009  INRIA Sophia-Antipolis (France).
 // All rights reserved.
 //
-// This file is part of CGAL (www.cgal.org); you can redistribute it and/or
-// modify it under the terms of the GNU Lesser General Public License as
-// published by the Free Software Foundation; either version 3 of the License,
-// or (at your option) any later version.
-//
-// Licensees holding a valid commercial license may use this file in
-// accordance with the commercial license agreement provided with the software.
-//
-// This file is provided AS IS with NO WARRANTY OF ANY KIND, INCLUDING THE
-// WARRANTY OF DESIGN, MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
+// This file is part of CGAL (www.cgal.org)
 //
 // $URL$
 // $Id$
-// SPDX-License-Identifier: LGPL-3.0+
+// SPDX-License-Identifier: LGPL-3.0-or-later OR LicenseRef-Commercial
 // 
 //
 // Author(s)     : Sebastien Loriot, Sylvain Pion
@@ -276,7 +267,7 @@ public:
         return Iso_rectangle_2();
       }
       
-      for (int i=0;i<static_cast<int>(static_cast<int>(get_IpePage()->count()));++i){
+      for (int i=0;i<static_cast<int>(get_IpePage()->count());++i){
         if (get_IpePage()->select(i)==ipe::ENotSelected)
           continue;
